+++ CHANGELOG.md
@@ -99,6 +99,9 @@
 - Added `Wizaplace\SDK\CreditCard\CreditCardService::getCreditCards`
 - Added `Wizaplace\SDK\CreditCard\CreditCardService::getCreditCard`
 - Added `Wizaplace\SDK\CreditCard\CreditCardService::getRegistrationUrl`
+- Added `Wizaplace\SDK\Catalog\ProductFilter::IS_SUBSCRIPTION`
+- Added `Wizaplace\SDK\Catalog\ProductFilter::getIsSubscription`
+- Added `Wizaplace\SDK\Catalog\ProductFilter::setIsSubscription`
 
 ## 1.80.0